extract precise-code-intel-worker to shared package for external invocation
This is a copy-and-paste of the enterprise/cmd/precise-code-intel-worker main func to a shared package so that it can be invoked from a different entrypoint. There are no logic changes.
