+++ enterprise/cmd/precise-code-intel-worker/shared/config.go
@@ -1,4 +1,4 @@
-package main
+package shared
 
 import (
 	"time"
